SBERDOMA-763 move register callback outside function
